\name{paircompviz-package}
\alias{paircompviz-package}
\alias{paircompviz}
\docType{package}
\title{Multiple comparison test visualization}
\description{
  This package provides visualization of the results from the multiple (i.e. pairwise)
  comparison tests such as pairwise.t.test, pairwise.prop.test or pairwise.wilcox.test. The groups
  being compared are visualized as nodes in Hasse diagram. Such approach enables very clear and
  vivid depiction of which group is significantly greater than which others, especially if comparing
  a large number of groups.
}
\details{
\tabular{ll}{
Package: \tab paircompviz\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-10-24\cr
License: \tab GPL-2\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Michal Burda

Maintainer: Michal Burda <michal.burda@osu.cz>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\keyword{ hplot }
\keyword{ htest }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
#~~ simple examples of the most important functions ~~
}
